Rewrite unused analyzer for multiple languages
Replace the previous monolithic unused-code scanner with a modular AnalyzerService and dedicated analyzers for C#, JavaScript, Dart, and Swift. Adds improved file collection and skip logic, richer regex-based symbol detection, cross-file analysis, pubspec dependency and asset checks for Dart, deduplication/sorting of results, and unit tests. Also includes a small fix in commands/mod.rs to remove redundant i64::from conversions when returning matcher scores.

+++ src/commands/mod.rs
@@ -564,15 +564,15 @@ fn rank_project_search_match(
     query: &str,
 ) -> Option<(u8, i64)> {
     if let Some(score) = matcher.fuzzy_match(&project.name, query) {
-        return Some((3, i64::from(score)));
+        return Some((3, score));
     }
 
     if let Some(score) = matcher.fuzzy_match(&project.category, query) {
-        return Some((2, i64::from(score)));
+        return Some((2, score));
     }
 
     if let Some(score) = matcher.fuzzy_match(&project.project_type, query) {
-        return Some((1, i64::from(score)));
+        return Some((1, score));
     }
 
     None
